+++ R/trip.r
@@ -1,0 +1,23 @@
+#' @export
+#' @rdname sptable
+map_table.trip <- function(x, ...) {
+  
+  if (!requireNamespace("trip", quietly = TRUE)) {
+    stop("trip package is required but not available\n try 'install.packages(\"trip\")'")
+  }
+  mpts <- as_trip_MultiPoints(x)
+  tabs <- map_table(mpts)
+  ## now put the stuff back on
+  tabs$v <- bind_cols(tabs$v, x@data)
+  tabs
+}
+
+as_trip_MultiPoints <- function(x, ...) {
+  tor <- x@TOR.columns
+  SpatialMultiPointsDataFrame(
+    ## sheesh
+    lapply(split(as.data.frame(coordinates(x)), x[[tor[2L]]]), as.matrix), 
+                              data.frame(object_ = seq_len(length(unique(x[[tor[2L]]]))), stringsAsFactors = FALSE), 
+                              proj4string = CRS(proj4string(x)), match.ID = FALSE)
+}
+